+++ src/views/Calendar.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { useState } from "react"
 import { Calendar as C, momentLocalizer } from 'react-big-calendar'
 import moment from 'moment'
 
@@ -13,18 +13,30 @@ import
 from "reactstrap";
 
 import "../assets/css/styles.scss"
+import Cookies from "js-cookie";
 
 
 const Calendar = () => {
-    const localizer = momentLocalizer(moment)
-    const myEventsList = [
-        {
-            id: 0,
-            title: "hey",
-            start: new Date(2021, 6,15),
-            end: new Date(2021, 6,19),
-        }
-    ]
+    const localizer = momentLocalizer(moment);
+
+    const [list, setList] = useState(null);
+
+    function OnLoad() { 
+        fetch('http://10.25.0.5:5454/getCalendar', {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json', 'token': Cookies.get("token")}
+        })
+        .then(res => res.json())
+        .then(res => {
+            if(res[0].error != 404)
+                setList(res)
+            else {
+                Cookies.remove("token");
+                window.location.href = "/";
+            } 
+        })
+      };
+
     const TimeElapsed = Date.now();
     const Today = new Date(TimeElapsed);
     const ColoredDateCellWrapper = ({ children }) =>
@@ -33,23 +45,27 @@ const Calendar = () => {
         backgroundColor: 'transparent',
         },
     })
-    return(
-        <>
-        {
-            console.log(Today)
-        }
+        return(
+        
+            list == null?  
+            <>
+                {OnLoad()}
+                Loading
+            </>
+            :
+            <>
             <div className="content">
                 <Row>
                     <Col xs="12">
                         <Card>
                              <CardHeader>
-                                 <labe>
+                                 <label>
                                      Task list
-                                 </labe>
+                                 </label>
                              </CardHeader>
                              <CardBody style={{height: 810}}>
                                 <C
-                                events={myEventsList}
+                                events={list}
                                 step={60}
                                 showMultiDayTimes
                                 defaultDate={Today}