+++ src/views/LoginDemo.js
@@ -5,25 +5,42 @@ import "../assets/css/default.scss";
 import "../components/FloatingLabel/FloatingLabel.css"
 
 const Login = () => {
-
     const OnMount = () => {
-        Cookies.set("isLogin", true);
-        window.location.href = "/user"
-    }
+        let username = document.getElementById("username").value;
+        let password = document.getElementById("password").value;
 
+        fetch('http://10.25.0.5:5454/login', {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json', 'user': username, 'pass': password}
+        })
+        .then(res => res.json())
+        .then(res => {
+            console.log(res);
+            if(res.success) {
+                Cookies.set("token", res.token);
+                setTimeout(() => {
+                    window.location.href = "/user";
+                }, 500);
+            }
+        })
+    }
     return(
         <>
+            {
+                window.location.href.split("/")[3] != "?#" ? window.location.href = "/?#"
+                : null
+            }
             <div className="body">
                 <div className="bg"/>
                 <div className="b-strip">
-                    <form className="f-log" target="/user" onSubmit={() => OnMount()}>
+                    <form className="f-log" action="#" onSubmit={() => OnMount()}>
                         <label>
                             Email address
                         </label>
                         <FloatingLabel 
                            label="test@test.com"
                            inputId="username"
-                           type="email"
+                           type="text"
                            required={true} 
                         />
                         <label>